fpibank-system/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
requests
asyncpg
pydantic[email]
httpx